+++ src/Plugin/BufferedAdd/BufferedAdd.php
@@ -12,18 +12,26 @@ namespace Solarium\Plugin\BufferedAdd;
 use Solarium\Core\Client\Endpoint;
 use Solarium\Core\Plugin\AbstractPlugin;
 use Solarium\Core\Query\DocumentInterface;
+use Solarium\Exception\RuntimeException;
+use Solarium\Plugin\BufferedAdd\Delete\Id as DeleteById;
+use Solarium\Plugin\BufferedAdd\Delete\Query as DeleteQuery;
+use Solarium\Plugin\BufferedAdd\Event\AddDeleteById as AddDeleteByIdEvent;
+use Solarium\Plugin\BufferedAdd\Event\AddDeleteQuery as AddDeleteQueryEvent;
 use Solarium\Plugin\BufferedAdd\Event\AddDocument as AddDocumentEvent;
 use Solarium\Plugin\BufferedAdd\Event\PostCommit as PostCommitEvent;
 use Solarium\Plugin\BufferedAdd\Event\PostFlush as PostFlushEvent;
 use Solarium\Plugin\BufferedAdd\Event\PreCommit as PreCommitEvent;
 use Solarium\Plugin\BufferedAdd\Event\PreFlush as PreFlushEvent;
+use Solarium\QueryType\Update\Query\Command\Add as AddCommand;
+use Solarium\QueryType\Update\Query\Command\Delete as DeleteCommand;
 use Solarium\QueryType\Update\Query\Query as UpdateQuery;
 use Solarium\QueryType\Update\Result as UpdateResult;
+use Symfony\Contracts\EventDispatcher\Event;
 
 /**
  * Buffered add plugin.
  *
- * If you need to add (or update) a big number of documents to Solr it's much more efficient to do so in batches.
+ * If you need to add (or update) and/or delete a big number of documents to Solr it's much more efficient to do so in batches.
  * This plugin makes this as easy as possible.
  */
 class BufferedAdd extends AbstractPlugin
@@ -45,14 +53,14 @@ class BufferedAdd extends AbstractPlugin
     protected $updateQuery;
 
     /**
-     * Buffered documents.
+     * Buffered documents to add, and/or document ids and/or queries to delete.
      *
-     * @var DocumentInterface[]
+     * @var (DocumentInterface|AbstractDelete)[]
      */
     protected $buffer = [];
 
     /**
-     * Set the endpoint for the documents.
+     * Set the endpoint for the documents and deletes.
      *
      * @param Endpoint $endpoint The endpoint to set
      *
@@ -172,14 +180,8 @@ class BufferedAdd extends AbstractPlugin
      */
     public function addDocument(DocumentInterface $document): self
     {
-        $this->buffer[] = $document;
-
         $event = new AddDocumentEvent($document);
-        $this->client->getEventDispatcher()->dispatch($event);
-
-        if (\count($this->buffer) === $this->options['buffersize']) {
-            $this->flush();
-        }
+        $this->addToBuffer($document, $event);
 
         return $this;
     }
@@ -187,7 +189,7 @@ class BufferedAdd extends AbstractPlugin
     /**
      * Add multiple documents.
      *
-     * @param array $documents
+     * @param DocumentInterface[] $documents
      *
      * @return self Provides fluent interface
      */
@@ -201,11 +203,75 @@ class BufferedAdd extends AbstractPlugin
     }
 
     /**
-     * Get all documents currently in the buffer.
+     * Add a document id to delete.
      *
-     * Any previously flushed documents will not be included!
+     * @param int|string $id
      *
-     * @return DocumentInterface[]
+     * @return self Provides fluent interface
+     */
+    public function addDeleteById($id): self
+    {
+        $delete = new DeleteById($id);
+        $event = new AddDeleteByIdEvent($delete);
+        $this->addToBuffer($delete, $event);
+
+        return $this;
+    }
+
+    /**
+     * Add multiple document ids to delete.
+     *
+     * @param (int|string)[] $ids
+     *
+     * @return self Provides fluent interface
+     */
+    public function addDeleteByIds(array $ids): self
+    {
+        foreach ($ids as $id) {
+            $this->addDeleteById($id);
+        }
+
+        return $this;
+    }
+
+    /**
+     * Add a query to delete matching documents.
+     *
+     * @param string $query
+     *
+     * @return self Provides fluent interface
+     */
+    public function addDeleteQuery(string $query): self
+    {
+        $delete = new DeleteQuery($query);
+        $event = new AddDeleteQueryEvent($delete);
+        $this->addToBuffer($delete, $event);
+
+        return $this;
+    }
+
+    /**
+     * Add multiple queries to delete matching documents.
+     *
+     * @param string[] $queries
+     *
+     * @return self Provides fluent interface
+     */
+    public function addDeleteQueries(array $queries): self
+    {
+        foreach ($queries as $query) {
+            $this->addDeleteQuery($query);
+        }
+
+        return $this;
+    }
+
+    /**
+     * Get all documents and deletes currently in the buffer.
+     *
+     * Any previously flushed documents and deletes will not be included!
+     *
+     * @return (DocumentInterface|AbstractDelete)[]
      */
     public function getDocuments(): array
     {
@@ -213,7 +279,7 @@ class BufferedAdd extends AbstractPlugin
     }
 
     /**
-     * Clear any buffered documents.
+     * Clear any buffered documents and deletes.
      *
      * @return self Provides fluent interface
      */
@@ -226,7 +292,7 @@ class BufferedAdd extends AbstractPlugin
     }
 
     /**
-     * Flush any buffered documents to Solr.
+     * Flush any buffered documents and deletes to Solr.
      *
      * @param bool|null $overwrite
      * @param int|null  $commitWithin
@@ -246,7 +312,7 @@ class BufferedAdd extends AbstractPlugin
         $event = new PreFlushEvent($this->buffer, $overwrite, $commitWithin);
         $this->client->getEventDispatcher()->dispatch($event);
 
-        $this->updateQuery->addDocuments($event->getBuffer(), $event->getOverwrite(), $event->getCommitWithin());
+        $this->addBufferToQuery($event->getBuffer(), $event->getOverwrite(), $event->getCommitWithin());
         $result = $this->client->update($this->updateQuery, $this->getEndpoint());
         $this->clear();
 
@@ -259,7 +325,7 @@ class BufferedAdd extends AbstractPlugin
     /**
      * Commit changes.
      *
-     * Any remaining documents in the buffer will also be flushed
+     * Any remaining documents and deletes in the buffer will also be flushed
      *
      * @param bool|null $overwrite
      * @param bool|null $softCommit
@@ -275,7 +341,7 @@ class BufferedAdd extends AbstractPlugin
         $event = new PreCommitEvent($this->buffer, $overwrite, $softCommit, $waitSearcher, $expungeDeletes);
         $this->client->getEventDispatcher()->dispatch($event);
 
-        $this->updateQuery->addDocuments($event->getBuffer(), $event->getOverwrite());
+        $this->addBufferToQuery($event->getBuffer(), $event->getOverwrite());
         $this->updateQuery->addCommit($event->getSoftCommit(), $event->getWaitSearcher(), $event->getExpungeDeletes());
         $result = $this->client->update($this->updateQuery, $this->getEndpoint());
         $this->clear();
@@ -295,5 +361,76 @@ class BufferedAdd extends AbstractPlugin
     protected function initPluginType(): void
     {
         $this->updateQuery = $this->client->createUpdate();
+    }
+
+    /**
+     * Add a document or delete to the buffer and determine if the buffer needs to be flushed.
+     *
+     * @param DocumentInterface|AbstractDelete $documentOrDelete Document or delete to add
+     * @param Event                            $postAddEvent     Event to trigger after adding but before flushing
+     */
+    protected function addToBuffer($documentOrDelete, Event $postAddEvent): void
+    {
+        $this->buffer[] = $documentOrDelete;
+
+        $this->client->getEventDispatcher()->dispatch($postAddEvent);
+
+        if (\count($this->buffer) === $this->options['buffersize']) {
+            $this->flush();
+        }
+    }
+
+    /**
+     * Add all documents and deletes from a buffer as commands to the update query.
+     *
+     * @param (DocumentInterface|AbstractDelete)[] $buffer
+     * @param bool|null                            $overwrite
+     * @param int|null                             $commitWithin
+     */
+    protected function addBufferToQuery(array $buffer, ?bool $overwrite = null, ?int $commitWithin = null): void
+    {
+        $it = new \ArrayIterator($buffer);
+
+        $add = new AddCommand();
+        $del = new DeleteCommand();
+
+        if (null !== $overwrite) {
+            $add->setOverwrite($overwrite);
+        }
+
+        if (null !== $commitWithin) {
+            $add->setCommitWithin($commitWithin);
+        }
+
+        while ($it->valid()) {
+            $docOrDel = $it->current();
+
+            switch (true) {
+                case is_a($docOrDel, $isA = DocumentInterface::class):
+                    $command = $add->addDocument($docOrDel);
+                    break;
+                case is_a($docOrDel, $isA = AbstractDelete::class):
+                    switch ($docOrDel->getType()) {
+                        case AbstractDelete::TYPE_ID:
+                            $command = $del->addId($docOrDel->getId());
+                            break;
+                        case AbstractDelete::TYPE_QUERY:
+                            $command = $del->addQuery($docOrDel->getQuery());
+                            break;
+                        default:
+                            throw new RuntimeException('Unsupported delete type in buffer');
+                    }
+                    break;
+                default:
+                    throw new RuntimeException('Unsupported type in buffer');
+            }
+
+            $it->next();
+
+            if (!$it->valid() || !is_a($it->current(), $isA)) {
+                $this->updateQuery->add(null, clone $command);
+                $command->clear();
+            }
+        }
     }
 }