+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [unreleased]
+### Added
+- Plugin\BufferedAdd\BufferedAdd::addDeleteById() and addDeleteByIds()
+- Plugin\BufferedAdd\BufferedAdd::addDeleteQuery() and addDeleteQueries()
+
+
 ## [6.2.1]
 ### Added
 - Possibility to set the context on an endpoint for SolrCloud instances with a non-default `hostContext` or Solr instances behind a reverse proxy, defaults to `solr` if omitted